% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{createObject}
\alias{createObject}
\title{Create a highSpaClone object.}
\usage{
createObject(
  counts,
  location,
  min_avg_expression = 0.01,
  min_gene_counts = 100,
  gene_order_file = NULL,
  annotations_file = NULL,
  project = ""
)
}
\arguments{
\item{counts}{Gene-by-cell raw count matrix.}

\item{location}{A data.frame or matrix of spatial coordinates for each cell, with three columns (cell.id, x, y). The row names of location must match with the column names of counts.}

\item{min_avg_expression}{Minimum average expression per gene.}

\item{min_gene_counts}{Minimum total counts per cell.}

\item{gene_order_file}{The gene order file, contains chromosome, start, and stop position for each gene.}

\item{annotations_file}{Cell type annotation information. Should contain two columns: cell.id and cell.label. Must be a data.frame, .csv or .txt file.}

\item{project}{The name of the project.}
}
\value{
Return a preliminary highSpaClone object.
}
\description{
Create a highSpaClone object.
}
