% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{class}
\name{highSpaClone-class}
\alias{highSpaClone-class}
\title{The highSpaClone Class}
\description{
The highSpaClone Class
}
\section{Slots}{

\describe{
\item{\code{counts.data}}{Gene-by-cell raw count matrix.}

\item{\code{location}}{The spatial coordinates of each cell.}

\item{\code{chr_pos}}{The chromosome position corresponding to each bin, used for subsequent CNA heatmap drawing.}

\item{\code{smoothed.data}}{Gene-by-cell smoothed count matrix by gene order.}

\item{\code{cna.data}}{Final CNA matrix. Dimension: cell-by-bin.}

\item{\code{cluster}}{Tumor/Subclone label for each cell.}

\item{\code{annotation}}{Cell type annotation information.}

\item{\code{gene_order}}{Gene ordering information, contains chromosome, start, and stop position for each gene.}

\item{\code{project}}{The name of the project.}
}}

