% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{spatialplot}
\alias{spatialplot}
\title{Spatial scatter plot of inferred subclones}
\usage{
  spatialplot(
    cna_obj,
    colors = c("#ebe5c2", "#D57358", "#8a508f", "#023047", "#E64B35", "#4DBBD5", "#00A087",
               "#3C5488", "#F39B7F", "#8491B4", "#91D1C2", "#DC0000", "#7E6148", "#B09C85"),
    point_size = 0.01,
    use_x_reverse = FALSE,
    use_coord_flip = FALSE,
    title = ""
  )
}
\arguments{
  \item{cna_obj}{highSpaClone object.}
  
  \item{colors}{Character vector of hex/color names used for clusters.}
  
  \item{point_size}{Numeric point size passed to \code{geom_point()}.}
  
  \item{use_coord_flip}{Logical; if \code{TRUE}, apply \code{ggplot2::coord_flip()}
    to swap axes after reversing x (default \code{FALSE}).}
  
  \item{use_x_reverse}{Logical; if \code{TRUE}, reverses the x-axis direction
    (i.e., flips the plot horizontally). This option is useful when the
    spatial coordinates are mirrored relative to the original histology image
    or when you want to align the orientation with external annotations.
    Default is \code{FALSE}.}
  
  \item{title}{Character plot title.}
}

\description{
  Draws a spatial scatter plot from \code{cna_obj@cluster}, coloring points by
  clone label.
}
\examples{
  \dontrun{
    p <- spatialplot(cna_obj, point_size = 0.01, use_coord_flip = T, use_x_reverse = T,
                     title = "Tumor subclones")
    print(p)
  }
  
}
