% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highSpaClone-package.R
\docType{package}
\name{highSpaClone-package}
\alias{highSpaClone}
\alias{highSpaClone-package}
\title{highSpaClone: Copy Number Variation Inference and Tumor Subclone Analysis for High-Resolution Spatial Transcriptomics}
\description{
highSpaClone is a computational framework for spatial CNV inference and tumor subclone classification from high-resolution spatial transcriptomics (SRT) data.
}
\author{
\strong{Maintainer}: Chenxuan Zang \email{CZang@mdanderson.org} (\href{https://orcid.org/0000-0001-5237-0351}{ORCID})

}
\keyword{internal}
